requirements.txt
streamlit>=1.33,<2
pandas>=2.2,<3
dropbox>=12,<13
sqlalchemy>=2.0,<3
psycopg2-binary>=2.9,<3
openpyxl>=3.0,<4
xlrd>=2.0,<3
fastapi>=0.115,<1
uvicorn>=0.30,<1
requests>=2.32,<3
rapidfuzz>=3.9,<4
beautifulsoup4>=4.12,<5